move shared test state
